requirements.txt
# Core Python & Data Science
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Geospatial Processing
rioxarray==0.15.1
rasterio==1.3.8
geopandas==0.13.2
pyproj==3.6.1
shapely==2.0.1

# Planetary Computer & STAC
pystac-client==0.7.3
planetary-computer==0.4.9
odc-stac==0.3.6
stackstac==0.4.2

# Machine Learning
xgboost==1.7.6
shap==0.42.1

# Deep Learning
torch==2.0.1
torchvision==0.15.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
ipython==8.14.0

# Utilities
tqdm==4.65.0
python-dotenv==1.0.0